{"types":"C","runs":16}

+++ problem16.java
@@ -3,10 +3,10 @@ import acm.program.ConsoleProgram;
 
 public class problem16 extends ConsoleProgram {
 	public void run() {
-		double firstNum = (int) (readInt("First number is: "));
-		double secondNum = (int) (readInt("Second number is: "));
-		double thirdNum = (int) (readInt("Third number is : "));
-		double average = 1 / (1 / firstNum + 1 / secondNum + 1 / thirdNum);
+		int firstNum = readInt("First number is: ");
+		int secondNum = readInt("Second number is: ");
+		int thirdNum = readInt("Third number is : ");
+		double average = 1 / (1.0 / firstNum + 1.0 / secondNum + 1.0 / thirdNum);
 		println("Average is: " + average);		
 	}
 }
